add debugging info
Sometimes the image doesn't boot and doesn't even call setup(), let's
add some debugging messages in setup() and loop(). It seems it mostly
happens when the flash is configured in QIO mode and the image is
larger than 444 kB. It seems that changing the code a little bit is
enough to fix it, including uncommenting this debug code. Maybe it's
a checksum issue in the code.